+++ cli/main.rs
@@ -18,6 +18,7 @@ mod npm;
 mod ops;
 mod proc_state;
 mod resolver;
+mod semver;
 mod standalone;
 mod tools;
 mod tsc;
